Refactoring - Changed class structure, raised gradle version, added source tests.
