+++ neatle/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.library'
 ext {
     PUBLISH_GROUP_ID = 'si.inova'
     PUBLISH_ARTIFACT_ID = 'neatle'
-    PUBLISH_VERSION = '0.8.1'
+    PUBLISH_VERSION = '0.9.0'
 }
 
 android {
@@ -14,7 +14,7 @@ android {
         minSdkVersion 19
         targetSdkVersion 25
         versionCode 1
-        versionName "0.8.1"
+        versionName "0.9.0"
     }
 
     lintOptions {
@@ -30,6 +30,8 @@ configurations {
 dependencies {
     compile 'com.android.support:support-annotations:25.3.1'
     javadocDeps 'com.android.support:support-annotations:25.3.1'
+
+    testCompile 'junit:junit:4.12'
 }
 
 tasks.whenTaskAdded { task ->
