+++ neatle/src/main/java/si/inova/neatle/monitor/ConnectionMonitor.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package si.inova.neatle;
+package si.inova.neatle.monitor;
 
 import android.bluetooth.BluetoothAdapter;
 import android.bluetooth.BluetoothDevice;
@@ -31,6 +31,10 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.os.Handler;
+import android.support.annotation.RestrictTo;
+
+import si.inova.neatle.Neatle;
+import si.inova.neatle.util.NeatleLogger;
 
 /**
  * Monitors the connection with a bluetooth device.
@@ -40,23 +44,22 @@ public class ConnectionMonitor {
     private static final long DEFAULT_RECONNECT_TIMEOUT = 2500L;
     private static final long MAX_RECONNECT_TIMEOUT = 60 * 1000L;
 
+    private final Context context;
     private final BluetoothDevice device;
     private final Handler handler;
+
     private boolean keepAlive;
+
     private Connection connection;
     private ConnHandler connectionHandler = new ConnHandler();
     private ConnectionStateListener connectionStateListener;
     private ReconnectRunnable reconnectRunnable = new ReconnectRunnable();
-    private Context context;
 
-    private final BroadcastReceiver mReceiver = new BroadcastReceiver() {
+    private final BroadcastReceiver receiver = new BroadcastReceiver() {
         @Override
         public void onReceive(Context context, Intent intent) {
-            final String action = intent.getAction();
-
-            if (action.equals(BluetoothAdapter.ACTION_STATE_CHANGED)) {
-                final int state = intent.getIntExtra(BluetoothAdapter.EXTRA_STATE,
-                        BluetoothAdapter.STATE_OFF);
+            if (BluetoothAdapter.ACTION_STATE_CHANGED.equals(intent.getAction())) {
+                int state = intent.getIntExtra(BluetoothAdapter.EXTRA_STATE, BluetoothAdapter.STATE_OFF);
 
                 if (state == BluetoothAdapter.STATE_ON) {
                     onBluetoothTurnedOn();
@@ -65,6 +68,7 @@ public class ConnectionMonitor {
         }
     };
 
+    @RestrictTo(RestrictTo.Scope.LIBRARY)
     public ConnectionMonitor(Context context, BluetoothDevice device) {
         if (device == null) {
             throw new IllegalArgumentException("Device cannot be null");
@@ -108,7 +112,7 @@ public class ConnectionMonitor {
         }
 
         IntentFilter filter = new IntentFilter(BluetoothAdapter.ACTION_STATE_CHANGED);
-        context.registerReceiver(mReceiver, filter);
+        context.registerReceiver(receiver, filter);
 
         connection = Neatle.getConnection(context, device);
         connection.addConnectionHandler(connectionHandler);
@@ -127,7 +131,7 @@ public class ConnectionMonitor {
             return;
         }
 
-        context.unregisterReceiver(mReceiver);
+        context.unregisterReceiver(receiver);
         handler.removeCallbacks(reconnectRunnable);
         if (connection != null) {
             connection.removeConnectionStateListener(connectionHandler);
@@ -152,6 +156,7 @@ public class ConnectionMonitor {
     }
 
     private class ConnHandler implements ConnectionHandler, ConnectionStateListener {
+
         private long reconnectTimeout = DEFAULT_RECONNECT_TIMEOUT;
 
         @Override
@@ -170,7 +175,6 @@ public class ConnectionMonitor {
                 NeatleLogger.d("Will try to reconnect after " + (reconnectTimeout / 1000) + " seconds");
                 handler.postDelayed(reconnectRunnable, reconnectTimeout);
                 reconnectTimeout = Math.min(reconnectTimeout * 2, MAX_RECONNECT_TIMEOUT);
-
             } else if (newState == BluetoothAdapter.STATE_CONNECTED) {
                 reconnectTimeout = DEFAULT_RECONNECT_TIMEOUT;
             }
