+++ neatle/src/main/java/si/inova/neatle/monitor/ConnectionStateListener.java
@@ -22,39 +22,9 @@
  * SOFTWARE.
  */
 
-package si.inova.neatle;
+package si.inova.neatle.monitor;
 
-public abstract class OperationObserver {
+public interface ConnectionStateListener {
 
-    /**
-     * Invoked when a command has finished successfully.
-     *
-     * @param op      the operation
-     * @param result  the result of the command
-     * @param results the result of all the operations
-     */
-    public void onCommandSuccess(Operation op, CommandResult result, OperationResults results) {
-    }
-
-    public void onCommandExecuting(Operation op, OperationResults results) {
-    }
-
-    /**
-     * Invoked when a command has finished either successfully or unsuccessfully.
-     *
-     * @param operation the operation
-     * @param result    the result of the command
-     * @param results   the result of all the operations
-     */
-    public void onCommandFinished(Operation operation, CommandResult result, OperationResults results) {
-    }
-
-    /**
-     * Invoked when an operation has finished either successfully or unsuccessfully.
-     *
-     * @param op      the operation
-     * @param results the result of all the operations
-     */
-    public void onOperationFinished(Operation op, OperationResults results) {
-    }
+    void onConnectionStateChanged(Connection connection, int newState);
 }
