+++ neatle/src/main/java/si/inova/neatle/operation/Operation.java
@@ -22,25 +22,13 @@
  * SOFTWARE.
  */
 
-package si.inova.neatle;
+package si.inova.neatle.operation;
 
-/**
- * Listener for connection events on a bluetooth LE device.
- *
- * @see Neatle#getConnection
- */
-public interface ConnectionHandler {
+public interface Operation {
 
-    int ON_IDLE_KEEP_ALIVE = 1;
-    int ON_IDLE_DISCONNECT = 0;
+    void execute();
 
-    /**
-     * Called when there is no active subscription or any pending operations. By default
-     * an idle connection will be disconnected. In case of multiple handler, ON_IDLE_KEEP_ALIVE
-     * will win over ON_IDLE_DISCONNECT.
-     *
-     * @param connection the connection of this event
-     * @return ON_IDLE_DISCONNECT if
-     */
-    int onConnectionIdle(Connection connection);
+    void cancel();
+
+    boolean isCanceled();
 }
