+++ neatle/src/main/java/si/inova/neatle/operation/OperationObserver.java
@@ -22,9 +22,39 @@
  * SOFTWARE.
  */
 
-package si.inova.neatle;
+package si.inova.neatle.operation;
 
-public interface ConnectionStateListener {
+public abstract class OperationObserver {
 
-    void onConnectionStateChanged(Connection connection, int newState);
+    /**
+     * Invoked when a command has finished successfully.
+     *
+     * @param op      the operation
+     * @param result  the result of the command
+     * @param results the result of all the operations
+     */
+    public void onCommandSuccess(Operation op, CommandResult result, OperationResults results) {
+    }
+
+    public void onCommandExecuting(Operation op, OperationResults results) {
+    }
+
+    /**
+     * Invoked when a command has finished either successfully or unsuccessfully.
+     *
+     * @param operation the operation
+     * @param result    the result of the command
+     * @param results   the result of all the operations
+     */
+    public void onCommandFinished(Operation operation, CommandResult result, OperationResults results) {
+    }
+
+    /**
+     * Invoked when an operation has finished either successfully or unsuccessfully.
+     *
+     * @param op      the operation
+     * @param results the result of all the operations
+     */
+    public void onOperationFinished(Operation op, OperationResults results) {
+    }
 }
