+++ neatle/src/main/java/si/inova/neatle/source/InputSource.java
@@ -22,11 +22,34 @@
  * SOFTWARE.
  */
 
-package si.inova.neatle;
+package si.inova.neatle.source;
+
+import java.io.IOException;
 
 /**
- * A marker interface telling neatle to retrieve the data from
- * a non EDT thread, for example from a file
+ * An input source that provides byte array data when writing to a BTLE device.
  */
-public interface AsyncInputSource extends InputSource {
+public interface InputSource {
+
+    /**
+     * Opens the stream source.
+     *
+     * @throws IOException if an IO error occurs
+     */
+    void open() throws IOException;
+
+    /**
+     * Reads next "chunk" from this stream.
+     *
+     * @return the number of bytes read or null if there is nothing more to read.
+     * @throws IOException if an IO error occurs
+     */
+    byte[] nextChunk() throws IOException;
+
+    /**
+     * Closes the stream source.
+     *
+     * @throws IOException if an IO error occurs
+     */
+    void close() throws IOException;
 }
